requirements.txt
alembic==1.12.0
aniso8601==9.0.1
asgiref==3.7.2
backports.zoneinfo==0.2.1;python_version<"3.9"
beautifulsoup4==4.12.2
blinker==1.6.3
bootstrap4==0.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.2.7
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-filter==23.3
django-jazzmin==2.6.0
django-rest-auth==0.9.5
django-softdelete==0.10.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dominate==2.8.0
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
stripe==7.6.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
visitor==0.1.3
werkzeug==3.0.0
zipp==3.17.0
